Add sampleStdev and sampleVariance methods. The stdev and variance functions, as exposed, continue to be population functions.
When the sample sizes are large enough, the values for sampleStdev and sampleVariance are very close. However, I was using a sample size of 6 and was seeing a large difference. Ultimately, I tracked this down to sample vs population standard deviations (and variances). This package uses stdev as a population functions where software such as postgres and applications such as google sheets use sample functions.

- Update the README with new functions
- Tests for sampleStdev and sampleVariance
- Expose 4 new functions: sampleStdev, populationStdev, sampleVariance, and populationVariance
- Alias stdev to populationStdev for compatibility
- Alias variance to populationVariance for compatibility
- Add test for populationStdev of only one thing
- Bump version to 2.2.0